move nav to component
create folder for users
