+++ app/containers/LoginOrRegister.jsx
@@ -2,7 +2,7 @@ import React, { Component, PropTypes } from 'react';
 import ReactDOM from 'react-dom';
 import classNames from 'classnames/bind';
 import { connect } from 'react-redux';
-import { manualLogin, signUp, toggleLoginMode } from 'actions/users';
+import { manualLogin, signUp, toggleLoginMode } from 'users/actions';
 import styles from 'css/components/login';
 import hourGlassSvg from 'images/hourglass.svg';
