+++ app/navigation/container.js
@@ -1,10 +1,10 @@
 import React, { PropTypes } from 'react';
 import { Link } from 'react-router';
 import { connect } from 'react-redux';
-import { logOut } from 'actions/users';
+import { logOut } from 'users/actions';
 
 import classNames from 'classnames/bind';
-import styles from 'css/components/navigation';
+import styles from './style.css';
 
 const cx = classNames.bind(styles);
