+++ app/navigation/index.js
@@ -1,0 +1,1 @@
+export Navigation from './container';
